# internal/pkg/presentation/api/api.go
package api

import (
	"context"
	"encoding/base64"
	"encoding/hex"
	"encoding/json"
	"errors"

	"io"
	"net/http"

	"github.com/diwise/iot-agent/assets/docs"
	"github.com/diwise/iot-agent/internal/pkg/application"
	"github.com/diwise/iot-agent/internal/pkg/application/facades"
	"github.com/diwise/iot-agent/internal/pkg/application/types"

	"github.com/diwise/iot-agent/pkg/lwm2m"
	"github.com/diwise/senml"
	"github.com/diwise/service-chassis/pkg/infrastructure/net/http/router"
	"github.com/diwise/service-chassis/pkg/infrastructure/o11y"
	"github.com/diwise/service-chassis/pkg/infrastructure/o11y/logging"
	"github.com/diwise/service-chassis/pkg/infrastructure/o11y/tracing"

	"go.opentelemetry.io/otel"
	"go.opentelemetry.io/otel/metric"
)

var tracer = otel.Tracer("iot-agent/api")

func RegisterHandlers(ctx context.Context, rootMux *http.ServeMux, app application.App, facade facades.EventFunc) error {
	const apiPrefix string = "/api/v0"

	docs.RegisterHandlers(ctx, rootMux)

	r := router.New(rootMux, router.WithPrefix(apiPrefix), router.WithTaggedRoutes(true))

	r.Post("/messages", NewIncomingMessageHandler(ctx, app, facade))
	r.Post("/messages/lwm2m", NewIncomingLWM2MMessageHandler(ctx, app))

	return nil
}

func NewIncomingMessageHandler(ctx context.Context, app application.App, facade facades.EventFunc) http.HandlerFunc {
	logger := logging.GetFromContext(ctx)

	return func(w http.ResponseWriter, r *http.Request) {
		var err error

		ctx, span := tracer.Start(r.Context(), "incoming-message")
		defer func() { tracing.RecordAnyErrorAndEndSpan(err, span) }()
		_, ctx, log := o11y.AddTraceIDToLoggerAndStoreInContext(span, logger, ctx)

		b, err := io.ReadAll(r.Body)
		defer r.Body.Close()
		if err != nil {
			log.Error("failed to read incoming message body", "err", err.Error())
			w.WriteHeader(http.StatusBadRequest)
			return
		}

		var im types.IncomingMessage
		err = json.Unmarshal(b, &im)
		if err != nil {
			log.Error("failed to unmarshal incoming message", "err", err.Error())
			w.WriteHeader(http.StatusBadRequest)
			return
		}

		evt, err := facade(ctx, im.Type, im.Data)
		if err != nil {
			log.Error("failed to decode sensor event using facade", "err", err.Error())
			w.WriteHeader(statusCodeForFacadeError(err))
			w.Write([]byte(err.Error()))
			return
		}

		if evt.DevEUI == "" {
			log.Debug("could not handle message", "type", im.Type, "reason", "DevEUI is missing")
			w.WriteHeader(http.StatusNoContent)
			return
		}

		// add source to event to use for auto create devices (TODO)
		evt.Source = im.Source

		err = app.HandleSensorEvent(ctx, evt)
		if err != nil {
			if !errors.Is(err, types.ErrNoDevice) {
				log.Error("failed to handle message", "err", err.Error())
			}

			w.WriteHeader(statusCodeForHandleSensorEventError(err))
			w.Write([]byte(err.Error()))

			return
		}

		w.WriteHeader(http.StatusCreated)
	}
}

func statusCodeForFacadeError(err error) int {
	var syntaxErr *json.SyntaxError
	var typeErr *json.UnmarshalTypeError
	var base64Err base64.CorruptInputError
	var hexErr hex.InvalidByteError

	switch {
	case errors.Is(err, types.ErrUnknownMessageType):
		return http.StatusUnprocessableEntity
	case errors.Is(err, types.ErrSensorIDMissing):
		return http.StatusUnprocessableEntity
	case errors.Is(err, types.ErrDecoderError):
		return http.StatusUnprocessableEntity
	case errors.As(err, &syntaxErr):
		return http.StatusUnprocessableEntity
	case errors.As(err, &typeErr):
		return http.StatusUnprocessableEntity
	case errors.As(err, &base64Err):
		return http.StatusUnprocessableEntity
	case errors.As(err, &hexErr):
		return http.StatusUnprocessableEntity
	default:
		return http.StatusInternalServerError
	}
}

func statusCodeForHandleSensorEventError(err error) int {
	switch {
	case errors.Is(err, types.ErrNoDevice):
		return http.StatusNotFound
	case errors.Is(err, types.ErrDecoderError):
		return http.StatusUnprocessableEntity
	default:
		return http.StatusInternalServerError
	}
}

func NewIncomingLWM2MMessageHandler(ctx context.Context, app application.App) http.HandlerFunc {
	messageCounter, err := otel.Meter("iot-agent/lwm2m").Int64Counter(
		"diwise.lwm2m.messages.total",
		metric.WithUnit("1"),
		metric.WithDescription("Total number of received lwm2m messages"),
	)

	logger := logging.GetFromContext(ctx)

	if err != nil {
		logger.Error("failed to create otel message counter", "err", err.Error())
	}

	return func(w http.ResponseWriter, r *http.Request) {
		var err error

		ctx, span := tracer.Start(r.Context(), "incoming-lwm2m-message")
		defer func() { tracing.RecordAnyErrorAndEndSpan(err, span) }()

		_, ctx, log := o11y.AddTraceIDToLoggerAndStoreInContext(span, logger, ctx)

		msg, err := io.ReadAll(r.Body)
		defer r.Body.Close()
		if err != nil {
			log.Error("failed to read incoming senML body", "err", err.Error())
			w.WriteHeader(http.StatusBadRequest)
			return
		}

		pack := senml.Pack{}

		err = json.Unmarshal(msg, &pack)
		if err == nil && len(pack) == 0 {
			err = errors.New("empty senML pack received")
		}

		if err != nil {
			log.Error("failed to decode incoming senML pack", "err", err.Error())
			w.WriteHeader(http.StatusBadRequest)
			return
		}

		messageCounter.Add(ctx, 1)

		deviceID := lwm2m.DeviceID(pack)
		err = app.HandleSensorMeasurementList(ctx, deviceID, pack)

		if err != nil {
			log.Error("failed to handle measurement list", "err", err.Error())
			w.WriteHeader(http.StatusInternalServerError)
			return
		}

		w.WriteHeader(http.StatusCreated)
	}
}